Refactor `Transaction` and `Validator` package
