+++ src/common/blockchain/validator.js
